+++ trading-card-backend/trading-cards-spring-boot/src/main/java/io/nology/tradingcardsspringboot/repositories/CardRepository.java
@@ -6,18 +6,18 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface CardRepository extends JpaRepository<Card, Long> {
+public interface CardRepository extends JpaRepository<Card, Integer> {
 
-    Optional<Card> findById(Long id);
+    Optional<Card> findById(int id);
     List<Card>findByName(String name);
-    List<Card>findByManaCostId(Long manaCostId);
-    List<Card>findByTypeLineId(Long typeLineId);
-    List<Card>findByExpansionSymbolId(Long expansionSymbolId);
+    List<Card>findByManaCostId(int manaCostId);
+    List<Card>findByTypeLineId(int typeLineId);
+    List<Card>findByExpansionSymbolId(int expansionSymbolId);
     List<Card>findByAbilities(String abilities);
     List<Card>findByFlavourText(String flavourText);
     List<Card>findBySymbolRarity(String symbolRarity);
     List<Card>findByArtistInfo(String artistInfo);
     List<Card>findByCollectorNum(int collectorNum);
     List<Card>findByPowerTough(String powerTough);
-    List<Card>findByCardBorderId(Long cardBorderId);
+    List<Card>findByCardBorderId(int cardBorderId);
 }
